Run the benchmark on the same ground the tests run on
The benchmark and the integration test both start real racer processes on real
block devices, and they had agreed on almost nothing about how. The test gave
each node a memfd behind a loop device, formatted it, mounted it, and tore the
whole thing down when the node dropped. The benchmark reloaded the brd module
to get its ram disks, which meant it could not run while anything else on the
host held one, reformatted whatever /dev/ram0 happened to be, and opened with a
pkill of every racer under its root. That is a benchmark that owns the machine
rather than its own storage, and it drifted from the test in the one place the
two should have been identical: the storage the numbers describe.

So unify the substrate on the test's. A memfd is per-run, needs no module
parameter, collides with nothing, and is released when the last descriptor
closes; it is still memory, and ext4 on a loop device is still a real filesystem
on a real block device. src/harness.rs now owns that setup and the racer process
lifecycle around it, spawning serve, reading the published metrics address and
device and fabric paths, installing a configuration atomically, waiting for the
reload, and shutting a set of processes down in the order their fabrics require.
The test and the benchmark share it and keep what genuinely differs: the test
still leaves the store absent so racer must place it, still grows and refuses to
shrink it, still rewires a partial mesh and replaces a member; the benchmark
still pins cores to their SMT siblings, reserves its stores up front, and runs
its ladder.

The cpu subcommand is gone. It timed crc32c and memcpy over a buffer resident in
L1 and reported a percentage of a core at a million IOPS, a number the end-to-end
run measures directly and under the cache pressure that makes it interesting.
Nothing consumed it. With it gone the two benchmark modules fit in src/bench.rs
in sections, and a bare racer-bench prints usage instead of starting a privileged
setup because the default subcommand no longer exists.
